pyproject.toml
[project]
name = "mcp-pdf-extract"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.11.0",
    "pdfplumber>=0.11.7",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
